Moved `tqdm.sklearn` into std.py and removed sklearn.py
